Clean up package for external consumption
Remove duplicate root-level source files (newer versions merged into
src/). Remove private flag, move react to peerDependencies, add files
field to limit installed contents to dist/.
