--- anta_helpers.ts
@@ -1,3 +1,0 @@
-export function hasChildren(children: React.ReactNode): boolean {
-  return Array.isArray(children) ? children.length > 0 : children != null
-}
