--- index.ts
@@ -1,2 +1,0 @@
-export { Progress } from './components/Progress'
-export { configure } from './jsx-runtime'
