+++ package.json
@@ -2,7 +2,7 @@
   "name": "anta",
   "version": "0.1.0",
   "type": "module",
-  "private": true,
+  "files": ["dist"],
   "exports": {
     ".": {
       "types": "./dist/index.d.ts",
@@ -34,8 +34,10 @@
     "typescript": "^5.7.0"
   },
   "dependencies": {
-    "classnames": "^2.5.0",
-    "react": "^19.0.0"
+    "classnames": "^2.5.0"
+  },
+  "peerDependencies": {
+    "react": "^18.0.0 || ^19.0.0"
   },
   "pnpm": {
     "onlyBuiltDependencies": [
