+++ src/elements/index.ts
@@ -3,7 +3,7 @@ import './a-progress.css'
 
 export { AProgressElement, register_a_progress } from './a-progress'
 
-// Auto-register in browser context
+// typeof guard: direct reference to undeclared variable throws ReferenceError, typeof does not
 if (typeof customElements !== 'undefined') {
   register_a_progress()
 }
