--- types.d.ts
@@ -1,5 +1,0 @@
-declare module '*.module.css' {
-  const classes: { [key: string]: string }
-  export default classes
-}
-
